Move LinkHandle
